Move NSQ protocol into a dedicated control module, and begin restructure:
Prompted by @mreiferson's recently published documentation, I'm going to break
provider and comsumer functionality into their own modules and make the
`Connection` an accessor for the two. It turns out that NSQ is transactional
for everything but comsumer-side messages (the push'd ones). This means that
those "OK" responses that I just sort of been ignoring are actually meaningful.